Sync fork

+++ web/reset/index.php
@@ -13,33 +13,41 @@ include($_SERVER['DOCUMENT_ROOT']."/inc/main.php");
 if ((!empty($_POST['user'])) && (empty($_POST['code']))) {
     $v_user = escapeshellarg($_POST['user']);
     $user = $_POST['user'];
+    $email = $_POST['email'];
     $cmd="/usr/bin/sudo /usr/local/hestia/bin/v-list-user";
     exec ($cmd." ".$v_user." json", $output, $return_var);
     if ( $return_var == 0 ) {
         $data = json_decode(implode('', $output), true);
-        $rkey = $data[$user]['RKEY'];
-        $fname = $data[$user]['FNAME'];
-        $lname = $data[$user]['LNAME'];
-        $contact = $data[$user]['CONTACT'];
-        $to = $data[$user]['CONTACT'];
-        $subject = __('MAIL_RESET_SUBJECT',date("Y-m-d H:i:s"));
-        $hostname = exec('hostname');
-        $from = __('MAIL_FROM',$hostname);
-        if (!empty($fname)) {
-            $mailtext = __('GREETINGS_GORDON_FREEMAN',$fname,$lname);
-        } else {
-            $mailtext = __('GREETINGS');
-        }
-        if (in_array(str_replace(':'.$_SERVER['SERVER_PORT'],'.conf',$_SERVER['HTTP_HOST']), array_merge(scandir('/etc/nginx/conf.d'),scandir('/etc/nginx/conf.d/domains'),scandir('/etc/apache2/conf.d/domains'),scandir('/etc/apache2/conf.d')))){
-            $mailtext .= __('PASSWORD_RESET_REQUEST',$_SERVER['HTTP_HOST'],$user,$rkey,$_SERVER['HTTP_HOST'],$user,$rkey);
-            if (!empty($rkey)) send_email($to, $subject, $mailtext, $from);
-            header("Location: /reset/?action=code&user=".$_POST['user']);
-            exit;
-        } else {
-            $ERROR = "<a class=\"error\">".__('Invalid host domain')."</a>";
-        }
-        unset($output);
+        if($email == $data[$user]['CONTACT']){
+            //genrate new rkey
+            exec ("/usr/bin/sudo /usr/local/hestia/bin/v-change-user-rkey ".$v_user."", $output, $return_var);
+            unset($output);
+            exec ($cmd." ".$v_user." json", $output, $return_var);
+            $data = json_decode(implode('', $output), true);
+            $rkey = $data[$user]['RKEY'];
+            $fname = $data[$user]['FNAME'];
+            $lname = $data[$user]['LNAME'];
+            $contact = $data[$user]['CONTACT'];
+            $to = $data[$user]['CONTACT'];
+            $subject = __('MAIL_RESET_SUBJECT',date("Y-m-d H:i:s"));
+            $hostname = exec('hostname');
+            $from = __('MAIL_FROM',$hostname);
+            if (!empty($fname)) {
+                $mailtext = __('GREETINGS_GORDON_FREEMAN',$fname,$lname);
+            } else {
+                $mailtext = __('GREETINGS');
+            }
+            if (in_array(str_replace(':'.$_SERVER['SERVER_PORT'],'.conf',$_SERVER['HTTP_HOST']), array_merge(scandir('/etc/nginx/conf.d'),scandir('/etc/nginx/conf.d/domains'),scandir('/etc/apache2/conf.d/domains'),scandir('/etc/apache2/conf.d')))){
+                $mailtext .= __('PASSWORD_RESET_REQUEST',$_SERVER['HTTP_HOST'],$user,$rkey,$_SERVER['HTTP_HOST'],$user,$rkey);
+                if (!empty($rkey)) send_email($to, $subject, $mailtext, $from);
+                header("Location: /reset/?action=code&user=".$_POST['user']);
+                exit;
+            } else {
+                $ERROR = "<a class=\"error\">".__('Invalid host domain')."</a>";
+            }
+        }     
     }
+    unset($output);
 }
 
 if ((!empty($_POST['user'])) && (!empty($_POST['code'])) && (!empty($_POST['password'])) ) {
@@ -52,24 +60,34 @@ if ((!empty($_POST['user'])) && (!empty($_POST['code'])) && (!empty($_POST['pass
             $data = json_decode(implode('', $output), true);
             $rkey = $data[$user]['RKEY'];
             if (hash_equals($rkey, $_POST['code'])) {
-                $v_password = tempnam("/tmp","vst");
-                $fp = fopen($v_password, "w");
-                fwrite($fp, $_POST['password']."\n");
-                fclose($fp);
-                $cmd="/usr/bin/sudo /usr/local/hestia/bin/v-change-user-password";
-                exec ($cmd." ".$v_user." ".$v_password, $output, $return_var);
-                unlink($v_password);
-                if ( $return_var > 0 ) {
-                    $ERROR = "<a class=\"error\">".__('An internal error occurred')."</a>";
-                } else {
-                    $_SESSION['user'] = $_POST['user'];
-                    header("Location: /");
-                    exit;
+                unset($output);
+                exec("/usr/bin/sudo /usr/local/hestia/bin/v-get-user-value ".$v_user." RKEYEXP", $output,$return_var);
+                if($output[0] > time() - 900){
+                    $v_password = tempnam("/tmp","vst");
+                    $fp = fopen($v_password, "w");
+                    fwrite($fp, $_POST['password']."\n");
+                    fclose($fp);
+                    $cmd="/usr/bin/sudo /usr/local/hestia/bin/v-change-user-password";
+                    exec ($cmd." ".$v_user." ".$v_password, $output, $return_var);
+                    unlink($v_password);
+                    if ( $return_var > 0 ) {
+                        sleep(5);
+                        $ERROR = "<a class=\"error\">".__('An internal error occurred')."</a>";
+                    } else {
+                        $_SESSION['user'] = $_POST['user'];
+                        header("Location: /");
+                        exit;
+                    }
+                }else{
+                    sleep(5);
+                    $ERROR = "<a class=\"error\">".__('Code has been expired')."</a>";
                 }
             } else {
+                sleep(5);
                 $ERROR = "<a class=\"error\">".__('Invalid username or code')."</a>";
             }
         } else {
+            sleep(5);
             $ERROR = "<a class=\"error\">".__('Invalid username or code')."</a>";
         }
     } else {
